+++ src/JavaScriptRuntime/Object.cs
@@ -4254,15 +4254,8 @@ namespace JavaScriptRuntime
             }
 
             // User-defined iterables: call obj[Symbol.iterator]().
-            object? iteratorMethod;
-            try
-            {
-                iteratorMethod = ObjectRuntime.GetItem(iterable, Symbol.iterator);
-            }
-            catch
-            {
-                iteratorMethod = null;
-            }
+            // GetIterator propagates abrupt completions from retrieving @@iterator.
+            object? iteratorMethod = ObjectRuntime.GetItem(iterable, Symbol.iterator);
 
             if (iteratorMethod is Delegate del)
             {
@@ -4291,15 +4284,14 @@ namespace JavaScriptRuntime
                 throw new JavaScriptRuntime.TypeError("Symbol.iterator is not a function");
             }
 
-            // Built-ins without an exposed Symbol.iterator surface fall back here.
-            if (iterable is string s)
-            {
-                return JavaScriptRuntime.String.CreateIterator(s);
-            }
-
             if (iterable is JavaScriptRuntime.Array arr)
             {
                 return arr.values();
+            }
+
+            if (iterable is string s)
+            {
+                return JavaScriptRuntime.String.CreateIterator(s);
             }
 
             if (iterable is ArgumentsObject argumentsObject)